services/shared/pyproject.toml
[project]
name = "yt-summarizer-shared"
version = "0.1.0"
description = "YT Summarizer Shared - Common code for API and workers"
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }
dependencies = [
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy>=2.0.0",
    "azure-identity>=1.19.0",
    "azure-storage-queue>=12.12.0",
    "azure-storage-blob>=12.23.0",
    "structlog>=24.4.0",
    "tenacity>=9.0.0",
    "alembic>=1.14.0",
    "pyodbc>=5.2.0",
    "aioodbc>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["shared"]

[tool.hatch.build.targets.wheel.sources]
"" = ""

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]
